refactor(sidecrush): rename healthcheck module to blockbuilding_healthcheck and update imports
