+++ crates/infra/sidecrush/src/bin/blockbuilding-sidecar.rs
@@ -1,4 +1,4 @@
-use sidecrush::healthcheck::{self, alloy_client::AlloyEthClient, BlockProductionHealthChecker, HealthcheckConfig, Node};
+use sidecrush::blockbuilding_healthcheck::{self, alloy_client::AlloyEthClient, BlockProductionHealthChecker, HealthcheckConfig, Node};
 use clap::Parser;
 use tracing::Level;
 use metrics_exporter_dogstatsd::DogStatsDBuilder;
